Bugfixes in analyzer.
